refactor(cli): split eval benchmark command into its own module
The eval command file exceeded the 500 LOC invariant; moves the
benchmark subcommand and its timestamp formatter to eval/benchmark.rs.
